a6/requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
reportlab>=4.0.0
